# Conflicts:
#	scopes/harmony/config/config.main.runtime.ts

+++ scopes/component/merging/merging.main.runtime.ts
@@ -8,12 +8,11 @@ import {
   getMergeStrategyInteractive,
   MergeOptions,
 } from '@teambit/legacy/dist/consumer/versions-ops/merge-version';
-import SnappingAspect, { SnapResults, SnappingMain, TagResults } from '@teambit/snapping';
+import SnappingAspect, { SnappingMain, TagResults } from '@teambit/snapping';
 import hasWildcard from '@teambit/legacy/dist/utils/string/has-wildcard';
 import mapSeries from 'p-map-series';
 import { ComponentID, ComponentIdList } from '@teambit/component-id';
 import { BitError } from '@teambit/bit-error';
-import GeneralError from '@teambit/legacy/dist/error/general-error';
 import { LaneId } from '@teambit/lane-id';
 import { AutoTagResult } from '@teambit/legacy/dist/scope/component-ops/auto-tag';
 import { UnmergedComponent } from '@teambit/legacy/dist/scope/lanes/unmerged-components';
@@ -48,6 +47,7 @@ import {
 } from '@teambit/config-merger';
 import { SnapsDistance } from '@teambit/legacy/dist/scope/component-ops/snaps-distance';
 import { InstallMain, InstallAspect } from '@teambit/install';
+import ScopeAspect, { ScopeMain } from '@teambit/scope';
 import { MergeCmd } from './merge-cmd';
 import { MergingAspect } from './merging.aspect';
 import { MergeStatusProvider, MergeStatusProviderOptions } from './merge-status-provider';
@@ -81,6 +81,13 @@ export type FailedComponents = { id: ComponentID; unchangedMessage: string; unch
 // fileName is PathLinux. TS doesn't let anything else in the keys other than string and number
 export type FilesStatus = { [fileName: string]: keyof typeof FileStatus };
 
+export type MergeSnapResults = {
+  snappedComponents: ConsumerComponent[];
+  autoSnappedResults: AutoTagResult[];
+  removedComponents?: ComponentIdList;
+  exportedIds?: ComponentID[];
+} | null;
+
 export type ApplyVersionResult = { id: ComponentID; filesStatus: FilesStatus };
 
 export type ApplyVersionResults = {
@@ -91,11 +98,7 @@ export type ApplyVersionResults = {
   addedComponents?: ComponentID[]; // relevant when restoreMissingComponents is true (e.g. bit lane merge-abort)
   resolvedComponents?: ConsumerComponent[]; // relevant for bit merge --resolve
   abortedComponents?: ApplyVersionResult[]; // relevant for bit merge --abort
-  mergeSnapResults?: {
-    snappedComponents: ConsumerComponent[];
-    autoSnappedResults: AutoTagResult[];
-    removedComponents?: ComponentIdList;
-  } | null;
+  mergeSnapResults?: MergeSnapResults;
   mergeSnapError?: Error;
   leftUnresolvedConflicts?: boolean;
   verbose?: boolean;
@@ -109,6 +112,7 @@ export type ApplyVersionResults = {
 export class MergingMain {
   constructor(
     private workspace: Workspace,
+    private scope: ScopeMain,
     private install: InstallMain,
     private snapping: SnappingMain,
     private checkout: CheckoutMain,
@@ -221,7 +225,8 @@ export class MergingMain {
     build: boolean;
     skipDependencyInstallation?: boolean;
   }): Promise<ApplyVersionResults> {
-    const consumer = this.workspace.consumer;
+    const consumer = this.workspace?.consumer;
+    const legacyScope = this.scope.legacyScope;
     const componentWithConflict = allComponentsStatus.find(
       (component) => component.mergeResults && component.mergeResults.hasConflicts
     );
@@ -252,8 +257,9 @@ export class MergingMain {
 
     const allConfigMerge = compact(succeededComponents.map((c) => c.configMergeResult));
 
-    const { workspaceDepsUpdates, workspaceDepsConflicts } =
-      await this.configMerger.updateWorkspaceJsoncWithDepsIfNeeded(allConfigMerge);
+    const { workspaceDepsUpdates, workspaceDepsConflicts } = this.workspace
+      ? await this.configMerger.updateWorkspaceJsoncWithDepsIfNeeded(allConfigMerge)
+      : { workspaceDepsUpdates: undefined, workspaceDepsConflicts: undefined };
 
     let workspaceConfigConflictWriteError: Error | undefined;
     if (workspaceDepsConflicts) {
@@ -261,17 +267,17 @@ export class MergingMain {
         workspaceDepsConflicts
       );
     }
-    await this.configMerger.generateConfigMergeConflictFileForAll(allConfigMerge);
+    if (this.workspace) await this.configMerger.generateConfigMergeConflictFileForAll(allConfigMerge);
 
-    if (currentLane) consumer.scope.objects.add(currentLane);
+    if (currentLane) legacyScope.objects.add(currentLane);
 
-    await consumer.scope.objects.persist(); // persist anyway, if currentLane is null it should save all main heads
+    await legacyScope.objects.persist(); // persist anyway, if currentLane is null it should save all main heads
 
-    await consumer.scope.objects.unmergedComponents.write();
+    await legacyScope.objects.unmergedComponents.write();
 
-    await consumer.writeBitMap(`merge ${otherLaneId.toString()}`);
+    if (this.workspace) await consumer.writeBitMap(`merge ${otherLaneId.toString()}`);
 
-    if (componentIdsToRemove.length) {
+    if (componentIdsToRemove.length && this.workspace) {
       const compBitIdsToRemove = ComponentIdList.fromArray(componentIdsToRemove);
       await deleteComponentsFiles(consumer, compBitIdsToRemove);
       await consumer.cleanFromBitMap(compBitIdsToRemove);
@@ -292,7 +298,7 @@ export class MergingMain {
       }
     }
 
-    const getSnapOrTagResults = async () => {
+    const getSnapOrTagResults = async (): Promise<MergeSnapResults> => {
       // if one of the component has conflict, don't snap-merge. otherwise, some of the components would be snap-merged
       // and some not. besides the fact that it could by mistake tag dependent, it's a confusing state. better not snap.
       if (noSnap || leftUnresolvedConflicts || componentsHasConfigMergeConflicts) {
@@ -305,16 +311,16 @@ export class MergingMain {
         const { taggedComponents, autoTaggedResults, removedComponents } = results;
         return { snappedComponents: taggedComponents, autoSnappedResults: autoTaggedResults, removedComponents };
       }
-      return this.snapResolvedComponents(consumer, snapMessage, build);
+      return this.snapResolvedComponents(snapMessage, build, currentLane?.toLaneId());
     };
-    let mergeSnapResults: ApplyVersionResults['mergeSnapResults'] = null;
+    let mergeSnapResults: MergeSnapResults = null;
     let mergeSnapError: Error | undefined;
-    const bitMapSnapshot = this.workspace.bitMap.takeSnapshot();
+    const bitMapSnapshot = this.workspace ? this.workspace.bitMap.takeSnapshot() : null;
     try {
       mergeSnapResults = await getSnapOrTagResults();
     } catch (err: any) {
       mergeSnapError = err;
-      this.workspace.bitMap.restoreFromSnapshot(bitMapSnapshot);
+      if (bitMapSnapshot) this.workspace.bitMap.restoreFromSnapshot(bitMapSnapshot);
     }
 
     return {
@@ -345,10 +351,11 @@ export class MergingMain {
     otherLane?: Lane | null // the lane we want to merged to our lane. (null if it's "main").
   ): Promise<ComponentMergeStatus[]> {
     const mergeStatusProvider = new MergeStatusProvider(
-      this.workspace,
+      this.scope,
       this.logger,
       this.importer,
       options,
+      this.workspace,
       currentLane || undefined,
       otherLane || undefined
     );
@@ -364,7 +371,7 @@ export class MergingMain {
     const componentsResults = await mapSeries(
       succeededComponents,
       async ({ currentComponent, id, mergeResults, resolvedUnrelated, configMergeResult }) => {
-        const modelComponent = await this.workspace.consumer.scope.getModelComponent(id);
+        const modelComponent = await this.scope.legacyScope.getModelComponent(id);
         const updatedLaneId = otherLaneId.isDefault() ? LaneId.from(otherLaneId.name, id.scope as string) : otherLaneId;
         return this.applyVersion({
           currentComponent,
@@ -380,16 +387,17 @@ export class MergingMain {
       }
     );
 
-    const compsToWrite = compact(componentsResults.map((c) => c.legacyCompToWrite));
-
-    const manyComponentsWriterOpts = {
-      consumer: this.workspace.consumer,
-      components: compsToWrite,
-      skipDependencyInstallation: true,
-      writeConfig: false, // @todo: should write if config exists before, needs to figure out how to do it.
-      reasonForBitmapChange: 'merge',
-    };
-    await this.componentWriter.writeMany(manyComponentsWriterOpts);
+    if (this.workspace) {
+      const compsToWrite = compact(componentsResults.map((c) => c.legacyCompToWrite));
+      const manyComponentsWriterOpts = {
+        consumer: this.workspace.consumer,
+        components: compsToWrite,
+        skipDependencyInstallation: true,
+        writeConfig: false, // @todo: should write if config exists before, needs to figure out how to do it.
+        reasonForBitmapChange: 'merge',
+      };
+      await this.componentWriter.writeMany(manyComponentsWriterOpts);
+    }
 
     return componentsResults;
   }
@@ -415,7 +423,7 @@ export class MergingMain {
     resolvedUnrelated?: ResolveUnrelatedData;
     configMergeResult?: ConfigMergeResult;
   }): Promise<ApplyVersionWithComps> {
-    const consumer = this.workspace.consumer;
+    const legacyScope = this.scope.legacyScope;
     let filesStatus = {};
     const unmergedComponent: UnmergedComponent = {
       id: { name: id.fullName, scope: id.scope },
@@ -424,7 +432,7 @@ export class MergingMain {
     };
     id = currentComponent ? currentComponent.id : id;
 
-    const modelComponent = await consumer.scope.getModelComponent(id);
+    const modelComponent = await legacyScope.getModelComponent(id);
 
     const addToCurrentLane = (head: Ref) => {
       if (!currentLane) throw new Error('currentLane must be defined when adding to the lane');
@@ -447,7 +455,7 @@ export class MergingMain {
         headOnCurrentLane: resolvedUnrelated.headOnCurrentLane,
         unrelatedLaneId: resolvedUnrelated.unrelatedLaneId,
       };
-      consumer.scope.objects.unmergedComponents.addEntry(unmergedComponent);
+      legacyScope.objects.unmergedComponents.addEntry(unmergedComponent);
       return { applyVersionResult: { id, filesStatus }, component: currentComponent, legacyCompToWrite };
     };
 
@@ -459,7 +467,7 @@ export class MergingMain {
     };
     if (mergeResults && mergeResults.hasConflicts && mergeStrategy === MergeOptions.ours) {
       markAllFilesAsUnchanged();
-      consumer.scope.objects.unmergedComponents.addEntry(unmergedComponent);
+      legacyScope.objects.unmergedComponents.addEntry(unmergedComponent);
       return { applyVersionResult: { id, filesStatus }, component: currentComponent || undefined };
     }
     if (resolvedUnrelated?.strategy === 'ours') {
@@ -468,7 +476,9 @@ export class MergingMain {
     }
     const remoteId = id.changeVersion(remoteHead.toString());
     const idToLoad = !mergeResults || mergeStrategy === MergeOptions.theirs ? remoteId : id;
-    const legacyComponent = await consumer.loadComponentFromModelImportIfNeeded(idToLoad);
+    const legacyComponent = this.workspace
+      ? await this.workspace.consumer.loadComponentFromModelImportIfNeeded(idToLoad)
+      : await legacyScope.getConsumerComponent(idToLoad); // when loading from the scope, we import all needed components first, so it should be fine. otherwise, change the code to import it here
     if (mergeResults && mergeStrategy === MergeOptions.theirs) {
       // in this case, we don't want to update .bitmap with the version of the remote. we want to keep the same version
       legacyComponent.version = id.version;
@@ -483,7 +493,7 @@ export class MergingMain {
       filesStatus = { ...filesStatus, ...modifiedStatus };
     }
 
-    await removeFilesIfNeeded(filesStatus, consumer, currentComponent || undefined);
+    if (this.workspace) await removeFilesIfNeeded(filesStatus, this.workspace.consumer, currentComponent || undefined);
 
     if (configMergeResult) {
       const successfullyMergedConfig = configMergeResult.getSuccessfullyMergedConfig();
@@ -500,7 +510,7 @@ export class MergingMain {
       if (mergeResults.hasConflicts && mergeStrategy === MergeOptions.manual) {
         unmergedComponent.unmergedPaths = mergeResults.modifiedFiles.filter((f) => f.conflict).map((f) => f.filePath);
       }
-      consumer.scope.objects.unmergedComponents.addEntry(unmergedComponent);
+      legacyScope.objects.unmergedComponents.addEntry(unmergedComponent);
     } else if (currentLane) {
       if (resolvedUnrelated) {
         // must be "theirs"
@@ -512,7 +522,7 @@ export class MergingMain {
       modelComponent.setHead(remoteHead);
       // mark it as local, otherwise, when importing this component from a remote, it'll override it.
       modelComponent.markVersionAsLocal(remoteHead.toString());
-      consumer.scope.objects.add(modelComponent);
+      legacyScope.objects.add(modelComponent);
     }
 
     return {
@@ -565,14 +575,25 @@ export class MergingMain {
   }
 
   private async snapResolvedComponents(
-    consumer: Consumer,
     snapMessage: string,
-    build: boolean
-  ): Promise<SnapResults | null> {
-    const unmergedComponents = consumer.scope.objects.unmergedComponents.getComponents();
+    build: boolean,
+    laneId?: LaneId
+  ): Promise<MergeSnapResults> {
+    const unmergedComponents = this.scope.legacyScope.objects.unmergedComponents.getComponents();
     this.logger.debug(`merge-snaps, snapResolvedComponents, total ${unmergedComponents.length.toString()} components`);
     if (!unmergedComponents.length) return null;
     const ids = ComponentIdList.fromArray(unmergedComponents.map((r) => ComponentID.fromObject(r.id)));
+    if (!this.workspace) {
+      const results = await this.snapping.snapFromScope(
+        ids.map((id) => ({ componentId: id.toString() })),
+        {
+          message: snapMessage,
+          build,
+          lane: laneId?.toString(),
+        }
+      );
+      return { ...results, autoSnappedResults: [] };
+    }
     return this.snapping.snap({
       legacyBitIds: ids,
       build,
@@ -603,13 +624,13 @@ export class MergingMain {
       componentIds.forEach((id) => {
         const entry = this.workspace.consumer.scope.objects.unmergedComponents.getEntry(id);
         if (!entry) {
-          throw new GeneralError(`unable to merge-resolve ${id.toString()}, it is not marked as unresolved`);
+          throw new BitError(`unable to merge-resolve ${id.toString()}, it is not marked as unresolved`);
         }
       });
       return componentIds;
     }
     const unresolvedComponents = this.workspace.consumer.scope.objects.unmergedComponents.getComponents();
-    if (!unresolvedComponents.length) throw new GeneralError(`all components are resolved already, nothing to do`);
+    if (!unresolvedComponents.length) throw new BitError(`all components are resolved already, nothing to do`);
     return unresolvedComponents.map((u) => ComponentID.fromObject(u.id));
   }
 
@@ -629,6 +650,7 @@ export class MergingMain {
   static dependencies = [
     CLIAspect,
     WorkspaceAspect,
+    ScopeAspect,
     SnappingAspect,
     CheckoutAspect,
     InstallAspect,
@@ -644,6 +666,7 @@ export class MergingMain {
   static async provider([
     cli,
     workspace,
+    scope,
     snapping,
     checkout,
     install,
@@ -657,6 +680,7 @@ export class MergingMain {
   ]: [
     CLIMain,
     Workspace,
+    ScopeMain,
     SnappingMain,
     CheckoutMain,
     InstallMain,
@@ -671,6 +695,7 @@ export class MergingMain {
     const logger = loggerMain.createLogger(MergingAspect.id);
     const merging = new MergingMain(
       workspace,
+      scope,
       install,
       snapping,
       checkout,
